frontend/requirements.txt
streamlit==1.35.0
requests==2.32.2
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
